Enables dry-run's client/server/none options with backwards compat
kubectl allows options [client|server|none] instead of just a boolean,
starting at some version I'm not sure of. This enables that behavior
with a fallback to if the user has passed true or false so folks don't
need to update their configurations.

I've mapped true->client and false->none but otherwise treating the
parameter as a string.